fix(nav): keep postponement href helpers out of client bundles
Header/MobileMenu imported pereiraTechDay, which pulls astro:content into
the Svelte island and failed production builds with ServerOnlyModule.
Move path helpers to a client-safe ptd-paths module.
