+++ src/components/layout/Header.svelte
@@ -9,7 +9,7 @@ import {
   stripLangPrefix,
 } from '@/lib/i18n';
 import { LANGUAGE_STORAGE_KEY } from '@/lib/language-preference';
-import { getPostponementAnnouncementHref } from '@/lib/pereiraTechDay';
+import { getPostponementAnnouncementHref } from '@/lib/ptd-paths';
 import { getTranslations } from '@/lib/translations';
 import MobileMenu from './MobileMenu.svelte';
 import SolidarityMark from './SolidarityMark.svelte';
